requirements.txt
git+https://github.com/xzzy/django_crispy_jcaptcha.git#egg=django_crispy_jcaptcha
asgiref==3.4.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.2.12
django-confy==1.0.4
django-countries==6.1.2
idna==2.9
pip==20.0.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2019.3
requests==2.26.0
setuptools==46.1.3
sqlparse==0.3.1
urllib3==1.26.5
wheel==0.34.2
whitenoise==5.3.0
django-crispy-forms==1.14.0
crypto-fernet==0.1.0
six>=1.10.0
dnspython==2.2.1
pyotp==2.6.0
